requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
anthropic>=0.86.0
openai>=2.30.0
python-dotenv==1.0.0
WTForms==3.1.1
